requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==5.0.4
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-countries==7.6.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-guardian==2.4.0
django-phonenumber-field==7.3.0
django-rest-auth==0.9.5
django-rest-framework-camel-case==2.0.1
django-rest-swagger==2.2.0
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
phonenumbers==8.13.35
phonenumberslite==8.13.35
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
